streamlit_ready_final/requirements.txt
streamlit==1.35.0
torch==1.12.1+cpu
torchvision==0.13.1+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html
transformers==4.30.2
Pillow==9.5.0
numpy==1.23.5
opencv-python-headless==4.7.0.72
matplotlib==3.7.1
